moved MetadiskClient to http module and renamed it to be Client.
